+++ tests/test_integrations/test_openai_agents/agents_app.py
@@ -1,135 +1,26 @@
-from agents import add_trace_processor
-import asyncio
-from deepeval.openai_agents import (
-    Runner,
-    Agent,
-    function_tool,
-    DeepEvalTracingProcessor,
-)
-
-from deepeval.tracing.context import update_current_trace
-from deepeval.prompt import Prompt
-
-add_trace_processor(DeepEvalTracingProcessor())
-
-prompt = Prompt(alias="asd")
-prompt.pull(version="00.00.01")
-
-
-@function_tool(metric_collection="test_collection_1")
-def get_current_weather(latitude: float, longitude: float) -> dict:
-    """
-    Fetches weather data for a given location using the Open-Meteo API.
-
-    Args:
-        latitude (float): The latitude of the location.
-        longitude (float): The longitude of the location.
-
-    Returns:
-        dict: A dictionary containing the weather data or error message.
-    """
-    # Return random dummy weather data for testing purposes
-    return {
-        "temperature_2m": 22.5,
-        "humidity": 55,
-        "apparent_temperature": 21.0,
-        "precipitation": 0.0,
-        "weather_code": 1,
-        "wind_speed_10m": 5.2,
-        "wind_direction_10m": 180,
-        "dummy": True,
-    }
-
-
-@function_tool(metric_collection="test_collection_1")
-def get_location_coordinates(city_name: str) -> dict:
-    """
-    Get latitude and longitude for a city name.
-
-    Args:
-        city_name (str): Name of the city
-
-    Returns:
-        dict: Dictionary with lat, lng coordinates
-    """
-    # Mock implementation - use real geocoding API in production
-    locations = {
-        "london": {"lat": 51.5074, "lng": -0.1278},
-        "tokyo": {"lat": 35.6762, "lng": 139.6503},
-        "new york": {"lat": 40.7128, "lng": -74.0060},
-    }
-
-    city_lower = city_name.lower()
-    if city_lower in locations:
-        return locations[city_lower]
-    return {"error": f"Location not found: {city_name}"}
-
-
-# Create the weather specialist agent
-weather_agent = Agent(
-    name="Weather Specialist Agent",
-    instructions="""
-    You are a weather agent. When providing current weather information 
-    (including temperature, humidity, wind speed/direction, precipitation, and weather codes), provide:
-    
-    1. A clear and concise summary of the weather conditions.
-    2. Practical suggestions or precautions for outdoor activities, health, or clothing based on the weather.
-    3. If severe weather is detected (e.g., heavy rain, thunderstorms, extreme temperatures), 
-       highlight necessary safety measures.
-    
-    Format your response in two sections:
-    Weather Summary:
-    - Briefly describe the weather in plain language.
-    
-    Suggestions:
-    - Offer actionable advice relevant to the weather conditions.
-    """,
-    tools=[get_location_coordinates, get_current_weather],
-    tool_use_behavior="run_llm_again",
-    # agent_metric_collection="test_collection_1",
-    llm_metric_collection="test_collection_1",
-    confident_prompt=prompt,
-)
-
-
-async def run_weather_agent(user_input: str):
-    """Run the weather agent with user input"""
-    runner = Runner()
-    result = await runner.run(
-        weather_agent,
-        user_input,
-        metric_collection="test_collection_1",
-        # name="test_name_1",
-        # user_id="test_user_id_1",
-        # thread_id="test_thread_id_1",
-        # tags=["test_tag_1"],
-        # metadata={"test_metadata_1": "test_metadata_1"},
-    )
-    return result.final_output
-
-
+from deepeval.openai_agents import Runner
+from deepeval.tracing import update_current_trace
 from agents import trace
-from multi_agents import triage_agent
+from multi_agent import triage_agent
+from weather_agent import weather_agent
 
-
-# with trace (group_id and metadata)
 async def main1():
+    # with trace (group_id and metadata)
     with trace(
-        workflow_name="test_workflow_1",
-        group_id="test_group_id_1",
-        metadata={"test_metadata_1": "test_metadata_1"},
+        workflow_name="test_workflow_1", # name of the trace,
+        group_id="test_group_id_1", # thread_id of the trace,
+        metadata={"test_metadata_1": "test_metadata_1"}, # metadata of the trace,
     ):
         user_query = "What's the weather like in London today?"
-        response_1 = await Runner.run(
+        response_1 = await Runner.run( # since thread id is present, input should not contain unwanted information.
             triage_agent,
             "Hola, ¿cómo estás?",
             metric_collection="test_collection_1",
-            thread_id="test",
         )
         response_2 = await Runner.run(
             weather_agent, user_query, metric_collection="test_collection_1"
         )
-        update_current_trace(input="initial input", output="final output")
+        # the input is of the first run and the output is of the second run.
 
 
 # without trace (group_id and metadata not present)
@@ -140,10 +31,10 @@ async def main2():
         "Hola, ¿cómo estás?",
         metric_collection="test_collection_1",
         thread_id="test",
-    )
+    ) # cleaned output
     response_2 = await Runner.run(
         weather_agent, user_query, metric_collection="test_collection_1"
-    )
+    ) # ouput will contain some unwanted information.
 
 
 async def main3():
@@ -153,9 +44,9 @@ async def main3():
         group_id="test_group_id_1",
         metadata={"test_metadata_1": "test_metadata_1"},
     ):
-        response_2 = await Runner.run(
-            weather_agent, user_query, metric_collection="test_collection_1"
-        )
+        response_2 = await Runner.run(weather_agent, user_query, metric_collection="test_collection_1")
+        update_current_trace(input="initial input", output="final output")
+    
     with trace(
         workflow_name="test_workflow_2",
         group_id="test_group_id_2",
@@ -180,8 +71,8 @@ async def main4():
             weather_agent, user_query, metric_collection="test_collection_1"
         )
         async for chunk in run_streamed_1.stream_events():
-            print(chunk, end="", flush=True)
-            print("=" * 50)
+            continue
+        
         run_streamed_2 = Runner.run_streamed(
             triage_agent,
             "Hola, ¿cómo estás?",
@@ -189,15 +80,11 @@ async def main4():
             thread_id="test",
         )
         async for chunk in run_streamed_2.stream_events():
-            print(chunk, end="", flush=True)
-            print("=" * 50)
+            continue
 
 
-def execute_agent():
-    asyncio.run(main1())
-    # asyncio.run(main2())
-    # asyncio.run(main3())
-    # asyncio.run(main4())
-
-
-# execute_agent()
+async def execute_agent():
+    await main1()
+    await main2()
+    await main3()
+    await main4()
